==> experiments_archive/515/PIH/3_split_train_test.ipynb <==
import pandas as pd
import numpy as np
import shutil
import os

# Load the cropped data CSV
cropped_data_path = './dataset/cropped/cropped_data.csv'
cropped_data = pd.read_csv(cropped_data_path)

# Separate features (image names) and labels (angles and radials)
X = cropped_data['Picture name']
y = cropped_data[['Angle [deg]', 'Radial distance [mm]']]

# Path to the cropped images
cropped_rgb_folder = './dataset/cropped/rgb/'
cropped_depth_folder = './dataset/cropped/depth/'

# Path to the train and test folders
train_rgb_folder = './dataset/train/cropped/rgb/'
train_depth_folder = './dataset/train/cropped/depth/'

test_rgb_folder = './dataset/test/cropped/rgb/'
test_depth_folder = './dataset/test/cropped/depth/'

# Split data into train and test sets
test_size = 1500
test_indices = np.random.choice(X.index, size=test_size, replace=False)
train_indices = X.index[~X.index.isin(test_indices)]

for index, row in cropped_data.iterrows():
    image_name = row['Picture name']
    angle = row['Angle [deg]']
    radial = row['Radial distance [mm]']
    
    source_rgb_path = os.path.join(cropped_rgb_folder, image_name)
    source_depth_path = os.path.join(cropped_depth_folder, image_name)
    
    if index in test_indices:
        target_rgb_path = os.path.join(test_rgb_folder, image_name)
        target_depth_path = os.path.join(test_depth_folder, image_name)
    else:
        target_rgb_path = os.path.join(train_rgb_folder, image_name)
        target_depth_path = os.path.join(train_depth_folder, image_name)
    
    shutil.copy(source_rgb_path, target_rgb_path)
    shutil.copy(source_depth_path, target_depth_path)

X_train, y_train = X[train_indices], y.iloc[train_indices]
X_test, y_test = X[test_indices], y.iloc[test_indices]

# Save train and test data to CSV
train_csv_path = './dataset/train/cropped/train_data.csv'
test_csv_path = './dataset/test/cropped/test_data.csv'

train_data = pd.concat([X_train, y_train], axis=1)
test_data = pd.concat([X_test, y_test], axis=1)

train_data.to_csv(train_csv_path, index=False)
test_data.to_csv(test_csv_path, index=False)
